requirements.txt
Django==1.5.1
Fabric==1.6.0
Markdown==2.3.1
South==0.7.6
distribute==0.6.31
dj-database-url==0.2.1
django-filter==0.6
djangorestframework==2.2.6
gunicorn==0.17.2
ipdb==0.7
ipython==0.13.2
mdx-smartypants==1.3
namedentities==1.301
paramiko==1.10.1
psycopg2==2.5
pycrypto==2.6
python-dateutil==2.1
python-memcached==1.48
pytz==2013b
readline==6.2.4.1
requests==1.2.0
six==1.3.0
wsgiref==0.1.2
xmltodict==0.4.6
